factor out filter custom label component
